+++ database/src/main/java/com/github/mukiva/weather_database/dao/ICurrentDao.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface ICurrentDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(current: CurrentDbo): Int
+    suspend fun insert(current: CurrentDbo): Long
     @Query("SELECT * FROM CurrentDbo WHERE id = :id")
     fun getById(id: Int): Flow<CurrentDbo>
     @Query("DELETE FROM CurrentDbo")
